Create async get anthropic, openai client funcs, move to reusable package
This package is where the get openai client functions also reside.
